login

+++ create_profile.php
@@ -44,6 +44,14 @@
     	}
   	}
   	
+  	function password_equals_confirm($password, $confirm_password) {
+  		if ($password==$confirm_password) {
+  			return true;
+  		} else {
+  			return false;
+  		}
+  	}
+  	
   	function num_users($db) {
   		$query = "select count(*) as c from User;";
   		$queryresult = $db->prepare($query);
@@ -55,6 +63,9 @@
     if($_POST["name"]) {
     	$email = htmlspecialchars($_POST["email"]);
     	$name = htmlspecialchars($_POST["name"]);
+    	$password = htmlspecialchars($_POST["password"]);
+    	echo "password: ".$password;
+    	$confirm_password = htmlspecialchars($_POST["confirm_password"]);
     	$doi = htmlspecialchars($_POST["doi"]);
     	$trunk_func = $_POST["trunk_func"];
     	$arms_func = $_POST["arms_func"];
@@ -65,13 +76,17 @@
     	$most_challenging = htmlspecialchars($_POST["most_challenging"]); 
     	$connect = htmlspecialchars($_POST["connect"]);    	
     	if( unique_email($email, $db) ) {
-    	  $id_num = num_users($db) + 1;
-		  $statement = 'INSERT INTO User VALUES ('.$id_num.', "'.$email.'", "'.$name.'", "'.$doi.'", '.$trunk_func.', '.$arms_func.', '.$fingers_func.', '.$legs_func.', "'.$injury_and_recovery.'", "'.$most_blissful.'", "'.$most_challenging.'", "'.$connect.'", 1);';
-	
-		  $dbcommand = $db->prepare($statement);
-		  $dbcommand->execute();
-		  
-		  echo "<center> Hello, ".$name.". You just created an account!.</center>"; 
+    		if (password_equals_confirm($password, $confirm_password) ) {
+			  $id_num = num_users($db) + 1;
+			  $statement = 'INSERT INTO User VALUES ('.$id_num.', "'.$email.'", "'.$name.'","'.sha1($password).'", "'.$doi.'", '.$trunk_func.', '.$arms_func.', '.$fingers_func.', '.$legs_func.', "'.$injury_and_recovery.'", "'.$most_blissful.'", "'.$most_challenging.'", "'.$connect.'", 1);';
+		
+			  $dbcommand = $db->prepare($statement);
+			  $dbcommand->execute();
+			  
+			  echo "<center> Hello, ".$name.". You just created an account!.</center>";     		
+    		} else {
+    			echo "<center> Your password and confirm password weren't the same!</center>";   
+    		}
 		} else {
 			echo "<center> Sorry, an account with that email already exists!</center>"; 
 		}	
@@ -81,7 +96,20 @@
   
   <form name="registration_form" method="POST" action="create_profile.php">
   <?php 
+	// filename: upload.form.php
+	// first let's set some variables
+	
+	// make a note of the current working directory relative to root.
+	$directory_self = str_replace(basename($_SERVER['PHP_SELF']), '', $_SERVER['PHP_SELF']);
+	
+	// make a note of the location of the upload handler
+	$uploadHandler = 'http://' . $_SERVER['HTTP_HOST'] . $directory_self . 'upload.processor.php';
+	
+	// set a max file size for the html upload form
+	$max_file_size = 10485760; // size in bytes			
     include ('./man_diagram.html');
+    
+    
   ?>
   </form>
